Add a layers!

+++ ScaleEngine-x64/src/Scale/Events/MouseEvent.h
@@ -6,10 +6,10 @@ namespace Scale {
 
 	class SCALE_API MouseMovedEvent : public Event {
 	public:
-		MouseMovedEvent(const float x, const float y) : m_MouseX(x), m_MouseY(y) {}
+		MouseMovedEvent(const double x, const double y) : m_MouseX(x), m_MouseY(y) {}
 
-		float GetX() const { return m_MouseX; }
-		float GetY() const { return m_MouseY; }
+		double GetX() const { return m_MouseX; }
+		double GetY() const { return m_MouseY; }
 
 		std::string toString() const override {
 			std::stringstream ss;
@@ -20,15 +20,15 @@ namespace Scale {
 		EVENT_CLASS_TYPE(MouseMoved)
 		EVENT_CLASS_CATEGORY(EC_Mouse | EC_Input)
 	private:
-		float m_MouseX, m_MouseY;
+		double m_MouseX, m_MouseY;
 	};
 
 	class SCALE_API MouseScrolledEvent : public Event {
 	public:
-		MouseScrolledEvent(const float xOffset, const float yOffset) : m_XOffset(xOffset), m_YOffset(yOffset) {}
+		MouseScrolledEvent(const double xOffset, const double yOffset) : m_XOffset(xOffset), m_YOffset(yOffset) {}
 
-		float GetXOffset() const { return m_XOffset; }
-		float GetYOffset() const { return m_YOffset; }
+		double GetXOffset() const { return m_XOffset; }
+		double GetYOffset() const { return m_YOffset; }
 
 		std::string toString() const override {
 			std::stringstream ss;
@@ -39,7 +39,7 @@ namespace Scale {
 		EVENT_CLASS_TYPE(MouseScrolled)
 		EVENT_CLASS_CATEGORY(EC_Mouse | EC_Input)
 	private:
-		float m_XOffset, m_YOffset;
+		double m_XOffset, m_YOffset;
 	};
 
 	class SCALE_API MouseButtonEvent : public Event
